Test and Fixed issues

+++ DCRM.Common/Dto/PatientseDto.cs
@@ -27,7 +27,7 @@ public partial class PatientseDto
 
     public string? Email { get; set; }
 
-    public sbyte? Age { get; set; }
+    public int? Age { get; set; }
 
     public int? Weight { get; set; }
